Insurance with storage-ROF table is implemented and working with new Matrix3D/2D classes.
Pseudo-ROF insurance has a bug related to the Matrix2D class. Something is not being freed properly, which is making the code to explode but, at the same time, Im not being able to free any memory either.

ReservoirExpansion class added, tested, and ready.

Quarry implemented but not tested. Simulation class is not sorting water sources by their IDs.